Version 2.0.0 - Migrate to JPA repository from legacy hibernate repository using hibernate criteria. - Implement ClickRepository, CampaignRepository, ActionRepository, PageviewRepository.

+++ src/main/java/com/indiepost/service/PostService.java
@@ -27,7 +27,7 @@ public interface PostService {
 
     List<PostSummaryDto> findByTagName(String tagName, int page, int maxResults, boolean isDesc);
 
-    List<PostSummaryDto> findTopRatedPosts(LocalDateTime since, LocalDateTime until, Long limit);
+    List<PostSummaryDto> findTopRatedPosts(LocalDateTime since, LocalDateTime until, Integer limit);
 
     List<PostSummaryDto> findScheduledPosts();
